+++ src-docs/App.vue
@@ -370,6 +370,7 @@ import issue64 from './json/issues/issue-64.json'
 import issue65 from './json/issues/issue-65.json'
 import issue33 from './json/issues/issue-33.json'
 import issue31 from './json/issues/issue-31.json'
+import issue56 from './json/issues/issue-56.json'
 import issue78 from './json/issues/issue-78.json'
 import issue79 from './json/issues/issue-79.json'
 import joditAsProp from './json/issues/jodit-as-prop.json'
@@ -573,6 +574,7 @@ export default {
           'issue/issue-44': issue44,
           'issue/issue-64': issue64,
           'issue/issue-65': issue65,
+          'issue/issue-56': issue56,
           'issue/issue-78': issue78,
           'issue/issue-79': issue79,
           'issue/jodit-as-prop': joditAsProp,